// src/app/MathPracticeApp/types.ts
// eslint-disable-next-line no-shadow,no-unused-vars
export enum Operations {
  // eslint-disable-next-line no-unused-vars
  Add = '+',
  // eslint-disable-next-line no-unused-vars
  Subtract = '-',
  // eslint-disable-next-line no-unused-vars
  Multiply = '*',
  // eslint-disable-next-line no-unused-vars
  Divide = '/'
}

export type operationsType = [Operations?, Operations?, Operations?, Operations?];